third_party/blink/web_tests/fast/html/imports/import-events-expected.txt

CONSOLE WARNING: HTML Imports is deprecated and will be removed in M73, around March 2019. Please use ES modules instead. See https://www.chromestatus.com/features/5144752345317376 for more details.
Test that events can be fired on imported documents.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.caughtFoo is ['foo@target', 'foo@root']
PASS successfullyParsed is true

TEST COMPLETE
